refactor: 도메인 패키지 구조 변경, 특강_선택_조회 메소드 리팩토링
