+++ src/main/java/com/hhplu/hhpluscleanarch/lecture/application/UserService.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhpluscleanarch.lecture.application;
 
-import com.hhplu.hhpluscleanarch.lecture.domain.LectureHistoryWithLecture;
+import com.hhplu.hhpluscleanarch.lecture.domain.dto.LectureHistoryWithLecture;
 import com.hhplu.hhpluscleanarch.lecture.exception.LectureHistoryNotFoundException;
 import com.hhplu.hhpluscleanarch.lecture.infrastructure.LectureHistoryRepository;
 import com.hhplu.hhpluscleanarch.lecture.infrastructure.UserHistoryResponse;
