+++ src/main/java/com/hhplu/hhpluscleanarch/lecture/infrastructure/LectureRepository.java
@@ -1,8 +1,8 @@
 package com.hhplu.hhpluscleanarch.lecture.infrastructure;
 
+import com.hhplu.hhpluscleanarch.lecture.common.LectureStatus;
 import com.hhplu.hhpluscleanarch.lecture.domain.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -10,7 +10,6 @@ import java.util.List;
 @Repository
 public interface LectureRepository extends JpaRepository<Lecture, Long> {
 
-    @Query("SELECT l FROM Lecture l WHERE l.capacity < 30 AND l.lectureStatus = com.hhplu.hhpluscleanarch.lecture.common.LectureStatus.OPENED")
-    List<Lecture> findAvailableLectures();
+    List<Lecture> findByCapacityLessThanAndLectureStatus(int capacity, LectureStatus lectureStatus);
 
 }
